backend/requirements.txt
Django>=5.1,<5.3
djangorestframework>=3.15
drf-spectacular>=0.27
celery[redis]>=5.4
redis>=5.2
psycopg[binary]>=3.2
dj-database-url>=2.2
python-dotenv>=1.0
pytest>=8.3
pytest-django>=4.9
ruff>=0.8
black>=24.10
